rename error handling container files
