+++ security/views/page_not_found_error.py
@@ -1,0 +1,18 @@
+from django.urls import reverse_lazy
+from django.shortcuts import (
+    HttpResponse,
+    render,
+    redirect,
+)
+
+def not_found_exception(request, exception):
+    return redirect('error_page')
+
+def not_found_page_view(request, *args, **kwargs):
+    context = {
+        'request': request,
+        'prevous_url': request. META,
+        'msg': kwargs,
+        'layout_path': 'layout/base.html',
+    }
+    return render(request, 'auth/pages/404.html', context)
